hybrid

+++ phylogeny/distTree_inc_hybrid.sh
@@ -10,8 +10,6 @@ if ($# != 3) then
 endif
 
 
-if (! -z $1/delete-hybrid)  exit 1
-
 
 if (! -e $1/hybrid) exit 0
 
@@ -38,19 +36,32 @@ if ("$3" == "") then
   exit 0
 endif
 
+
+if (! -z $1/delete-hybrid)  exit 1
+
 setMinus $3 $1/outlier.add > $1/test.list
 if ($?) exit 1
 
 rm $1/outlier.add
+
 
 # hybrid-add
 mkdir $1/hybrid
 if ($?) exit 1
 mkdir $1/log
 if ($?) exit 1
-# grid ??
-trav $1/test.list "distTree_inc_new2hybrid.sh $1 %f $1/hybrid/%f $1/log/%f"
-if ($?) exit 1
+
+set N = `wc -l $1/test.list`
+if ($N[1] > 20) then  # PAR
+	trav  -step 1  $1/test.list "$QSUB_5 -N j%n %QdistTree_inc_new2hybrid.sh $1 %f $1/hybrid/%f $1/log/%f%Q > /dev/null" 
+	if ($?) exit 1  
+	qstat_wait.sh 1
+	if ($?) exit 1
+else
+	trav -step 1  $1/test.list "distTree_inc_new2hybrid.sh $1 %f $1/hybrid/%f $1/log/%f"
+	if ($?) exit 1
+endif
+
 rmdir $1/log
 if ($?) exit 1
 trav -noprogress $1/hybrid "cat $1/hybrid/%f" > $1/hybrid-add
@@ -60,19 +71,23 @@ if ($?) exit 1
 
 rm $1/test.list
 
+
 # Cf. processing $1/hybrid
-wc -l $1/hybrid-add
-$1/hybrid2db.sh $1/hybrid-add
-if ($?) exit 1
 cut -f 1 $1/hybrid-add | sort > $1/delete-hybrid
 if ($?) exit 1
-$1/objects_in_tree.sh $1/delete-hybrid 0
-if ($?) exit 1
-trav -noprogress $1/delete-hybrid "cp /dev/null $1/outlier/%f"
-if ($?) exit 1
-mv $1/hybrid-add $1/hist/hybrid-add.$2
-if ($?) exit 1
-
+if (-z $1/hybrid-add) then
+  rm $1/hybrid-add
+else
+	wc -l $1/hybrid-add
+	$1/hybrid2db.sh $1/hybrid-add
+	if ($?) exit 1
+	$1/objects_in_tree.sh $1/delete-hybrid 0
+	if ($?) exit 1
+	trav -noprogress $1/delete-hybrid "cp /dev/null $1/outlier/%f"
+	if ($?) exit 1
+	mv $1/hybrid-add $1/hist/hybrid-add.$2
+	if ($?) exit 1
+endif
